update-patches.sh: always delete old Maildir
Putting the rm -rf Maildir towards the end of the script doesn't work if
an earlier command fails and the script exits.  Each time the script is
called by the cronjob but fails, the Maildir grows until all inodes are
exhausted and the file system is full.

Be defensive and always rm -rf Maildir before creating it.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>